move RegionFactory to factories module
