+++ addons/osfstorage/tests/factories.py
@@ -1,8 +1,7 @@
 #!/usr/bin/env python
 # encoding: utf-8
-from django.apps import apps
 from django.utils import timezone
-from factory import SubFactory, post_generation, Sequence
+from factory import SubFactory, post_generation
 from factory.django import DjangoModelFactory
 
 from osf_tests.factories import AuthUserFactory
@@ -10,30 +9,10 @@ from osf_tests.factories import AuthUserFactory
 from osf import models
 from addons.osfstorage.models import Region
 
-
-settings = apps.get_app_config('addons_osfstorage')
-
-
 generic_location = {
     'service': 'cloud',
     settings.WATERBUTLER_RESOURCE: 'resource',
     'object': '1615307',
-}
-
-generic_waterbutler_settings = {
-    'storage': {
-        'provider': 'glowcloud',
-        'container': 'osf_storage',
-        'use_public': True,
-    }
-}
-
-generic_waterbutler_credentials = {
-    'storage': {
-        'region': 'PartsUnknown',
-        'username': 'mankind',
-        'token': 'heresmrsocko'
-    }
 }
 
 
@@ -51,14 +30,3 @@ class FileVersionFactory(DjangoModelFactory):
         if not create:
             return
         self.reload()
-
-
-class RegionFactory(DjangoModelFactory):
-    class Meta:
-        model = Region
-
-    name = Sequence(lambda n: 'Region {0}'.format(n))
-    _id = Sequence(lambda n: 'us_east_{0}'.format(n))
-    waterbutler_credentials = generic_waterbutler_credentials
-    waterbutler_settings = generic_waterbutler_settings
-    waterbutler_url = 'http://123.456.test.woo'
